requirements.txt
APScheduler==3.11.0babel==2.16.0bcrypt==4.2.0blinker==1.7.0certifi==2024.2.2charset-normalizer==3.3.2click==8.1.7colorama==0.4.6contourpy==1.2.1cycler==0.12.1dnspython==2.7.0email-validator==2.2.0et-xmlfile==2.0.0filelock==3.13.4Flask==3.0.2flask-babel==4.0.0Flask-Bcrypt==1.0.1Flask-Cors==5.0.0Flask-Login==0.6.3Flask-Mail==0.10.0Flask-SQLAlchemy==3.1.1Flask-Table==0.5.0Flask-WTF==1.2.2fonttools==4.51.0fsspec==2024.3.1greenlet==3.1.1gunicorn==23.0.0idna==3.7itsdangerous==2.1.2Jinja2==3.1.3joblib==1.4.0kiwisolver==1.4.5llvmlite==0.43.0MarkupSafe==2.1.5matplotlib==3.8.4mpmath==1.3.0networkx==2.8.8numba==0.60.0numpy==1.26.4opencv-python==4.9.0.80openpyxl==3.1.5packaging==24.0pandas==2.2.2pillow==10.3.0psutil==5.9.8py-cpuinfo==9.0.0pynndescent==0.5.13pyodbc==5.2.0pyparsing==3.1.2python-dateutil==2.9.0.post0pytz==2024.1PyYAML==6.0.1requests==2.31.0scikit-learn==1.4.2scikit-surprise==1.1.4scipy==1.13.0seaborn==0.13.2six==1.16.0SQLAlchemy==2.0.36sympy==1.12thop==0.1.1.post2209072238threadpoolctl==3.4.0torch==2.2.2torchvision==0.17.2tqdm==4.66.2typing_extensions==4.11.0tzdata==2024.1tzlocal==5.3.1ultralytics==8.2.2umap-learn==0.5.7urllib3==2.2.1Werkzeug==3.0.1WTForms==3.2.1